Fix issue-9: criando testes unitarios  e algumas correções

+++ src/main/java/com/api/produtos/controller/ProdutoController.java
@@ -31,7 +31,7 @@ public class ProdutoController {
 	public Page<Produto> findAll(@RequestParam(required = false, defaultValue = "0") int page,
 								 @RequestParam(required = false, defaultValue = Integer.MAX_VALUE + "")	int size,
 								 @RequestParam(required = false) String nome,
-								 @RequestParam(required = false) BigDecimal quantidade){
+								 @RequestParam(required = false) int quantidade){
 		return produtoService.findAll(page, size, nome, quantidade);
 	}
