+++ src/main/java/com/api/produtos/model/Produto.java
@@ -25,7 +25,7 @@ public class Produto implements Serializable {
 	private String nome;
 
 	@NotNull
-	private BigDecimal quantidade;
+	private int quantidade;
 
 	@NotNull
 	private BigDecimal valor;
@@ -48,11 +48,11 @@ public class Produto implements Serializable {
 		this.nome = nome;
 	}
 
-	public BigDecimal getQuantidade() {
+	public int getQuantidade() {
 		return quantidade;
 	}
 
-	public void setQuantidade(BigDecimal quantidade) {
+	public void setQuantidade(int quantidade) {
 		this.quantidade = quantidade;
 	}
